Adjusting types and references to be long

+++ Source/Zencoder/ZencoderServices.cs
@@ -204,7 +204,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="jobId">The ID of the job to cancel.</param>
         /// <returns>The call response.</returns>
-        public CancelJobResponse CancelJob(int jobId)
+        public CancelJobResponse CancelJob(long jobId)
         {
             CancelJobRequest request = new CancelJobRequest(this)
             {
@@ -219,7 +219,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="jobId">The ID of the job to cancel.</param>
         /// <param name="callback">The call response.</param>
-        public void CancelJob(int jobId, Action<CancelJobResponse> callback)
+        public void CancelJob(long jobId, Action<CancelJobResponse> callback)
         {
             CancelJobRequest request = new CancelJobRequest(this)
             {
@@ -320,7 +320,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="jobId">The ID of the job to delete.</param>
         /// <returns>The call response.</returns>
-        public DeleteJobResponse DeleteJob(int jobId)
+        public DeleteJobResponse DeleteJob(long jobId)
         {
             DeleteJobRequest request = new DeleteJobRequest(this)
             {
@@ -335,7 +335,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="jobId">The ID of the job to delete.</param>
         /// <param name="callback">The call response.</param>
-        public void DeleteJob(int jobId, Action<DeleteJobResponse> callback)
+        public void DeleteJob(long jobId, Action<DeleteJobResponse> callback)
         {
             DeleteJobRequest request = new DeleteJobRequest(this)
             {
@@ -350,7 +350,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="jobId">The ID of the job to get details for.</param>
         /// <returns>The call response.</returns>
-        public JobDetailsResponse JobDetails(int jobId)
+        public JobDetailsResponse JobDetails(long jobId)
         {
             JobDetailsRequest request = new JobDetailsRequest(this)
             {
@@ -365,7 +365,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="jobId">The ID of the job to get details for.</param>
         /// <param name="callback">The call response.</param>
-        public void JobDetails(int jobId, Action<JobDetailsResponse> callback)
+        public void JobDetails(long jobId, Action<JobDetailsResponse> callback)
         {
             JobDetailsRequest request = new JobDetailsRequest(this)
             {
@@ -380,7 +380,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="outputId">The ID of the output (NOT the job ID) to get progress for.</param>
         /// <returns>The call response.</returns>
-        public JobProgressResponse JobProgress(int outputId)
+        public JobProgressResponse JobProgress(long outputId)
         {
             JobProgressRequest request = new JobProgressRequest(this)
             {
@@ -395,7 +395,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="outputId">The ID of the output (NOT the job ID) to get progress for.</param>
         /// <param name="callback">The call response.</param>
-        public void JobProgress(int outputId, Action<JobProgressResponse> callback)
+        public void JobProgress(long outputId, Action<JobProgressResponse> callback)
         {
             JobProgressRequest request = new JobProgressRequest(this)
             {
@@ -452,7 +452,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="jobId">The ID of the job to resubmit.</param>
         /// <returns>The call response.</returns>
-        public ResubmitJobResponse ResubmitJob(int jobId)
+        public ResubmitJobResponse ResubmitJob(long jobId)
         {
             ResubmitJobRequest request = new ResubmitJobRequest(this)
             {
@@ -467,7 +467,7 @@ namespace Zencoder
         /// </summary>
         /// <param name="jobId">The ID of the job to resubmit.</param>
         /// <param name="callback">The call response.</param>
-        public void ResubmitJob(int jobId, Action<ResubmitJobResponse> callback)
+        public void ResubmitJob(long jobId, Action<ResubmitJobResponse> callback)
         {
             ResubmitJobRequest request = new ResubmitJobRequest(this)
             {
